worker_ip bigint

+++ lib/schema.sql
@@ -62,7 +62,7 @@ CREATE TABLE "tests" (
   "skip_build"  boolean         NOT NULL,
   "started"     timestamptz,
   "finished"    timestamptz,
-  "worker_ip"   integer         NOT NULL DEFAULT 0,
+  "worker_ip"   bigint         NOT NULL DEFAULT 0,
   "worker_hostname"   varchar,
   -- Denormalised duplicate of runs.branch to speed up history search queries.
   "branch"      varchar         NOT NULL,
